Moved GetStorageInfo method to CommonUtils
