+++ Registry.Ports/ObjectSystem/IObjectSystem.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using Registry.Common.Model;
 using Registry.Ports.ObjectSystem.Model;
 
 namespace Registry.Ports.ObjectSystem
